arrivals and departures information showing up in tables!
And the tables do filtering, too! By last name or Stay ID.
And also crafted API call that gets all rooms. This will be what I need for changing a room's status from clean to dirty and/or dirty to clean. If a room is dirty it cannot be blocked.