saving this form is more of a pain than I thought
currently, it's mad about the date object formatting. the format of the date being saved needs to match the format in the database, but typeorm doesn't like the 000Z format for some reason. Going to make a date object in Angular to store the proper date in to fix that. Hopefully.